remove outdate JanusBlueprintsGraph
